turn into npm package usable in yari, add package + stage deploy github actions
