+++ .github/workflows/deploy-stage-api.yml
@@ -1,0 +1,28 @@
+name: Deploy bcd-utils-api data to s3
+
+jobs:
+  deploy-stage-api:
+    name: Deploy to stage
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: api
+
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          registry-url: "https://registry.npmjs.org/"
+          cache: npm
+      - run: npm ci
+      - run: npm update @mdn/browser-compat-data
+      - run: npm run generate
+      - name: Configure AWS credentials for stage
+        uses: aws-actions/configure-aws-credentials@v1
+        with:
+          aws-access-key-id: ${{ secrets.STAGE_AWS_ACCESS_KEY_ID }}
+          aws-secret-access-key: ${{ secrets.STAGE_AWS_ACCESS_KEY_SECRET }}
+          aws-region: us-west-2
+      - run: aws s3 sync --acl public-read out/ s3://mdn-content-stage/main/bcd/out/
+      - run: aws s3 sync --acl public-read --delete out/v0/current s3://mdn-content-stage/main/bcd/out/v0/current
