+++ .github/workflows/publish-package-api.yml
@@ -1,0 +1,29 @@
+name: Publish bcd-utils-api package to npm
+
+on:
+  release:
+    types:
+      - published
+
+jobs:
+  publish-package-api:
+    name: Publish package
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: api
+
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          registry-url: "https://registry.npmjs.org/"
+          cache: npm
+      - run: npm ci
+      - run: npm update @mdn/browser-compat-data
+      - run: npm test
+      - run: npm run build
+      - run: npm publish --access public
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
