+++ api/README.md
@@ -2,6 +2,6 @@
 
 Converts bcd data into json files usable by yari's bcd table.
 
-Run with `npm start`.
+Run with `npm run generate`.
 
-Outputs files into `./build`.
+Outputs files into `./out`.
