+++ api/tsconfig.json
@@ -1,7 +1,13 @@
 {
   "compilerOptions": {
+    "target": "ES2020",
+    "module": "CommonJS",
+    "moduleResolution": "node",
     "resolveJsonModule": true,
     "esModuleInterop": true,
-    "strict": true
-  }
+    "strict": true,
+    "outDir": "dist",
+    "declaration": true
+  },
+  "files": ["index.ts"]
 }
